tests/requirements.txt
pytest-asyncio==0.26.0
pytest-playwright==0.7.0
